Combine the admin settings into one JS file so we watch less resources
